+++ src/routes/v1/auth.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const AuthController = require('../controllers/AuthController');
+const AuthController = require('../../controllers/v1/AuthController');
 
 router.post('/', AuthController.login);
