+++ src/routes/v1/status.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const StatusController = require('../controllers/StatusController');
+const StatusController = require('../../controllers/v1/StatusController');
 
 router.get('/', StatusController.showStatus);
