+++ src/routes/v1/tutorials.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
-const TutorialController = require('../controllers/TutorialController');
-const jwtValidation = require('../middlewares/jwtValidation');
+const TutorialController = require('../../controllers/v1/TutorialController');
+const jwtValidation = require('../../middlewares/jwtValidation');
 
 router.use(jwtValidation);
