+++ src/tests/v1/Status.test.js
@@ -1,13 +1,13 @@
 const request = require('supertest');
 const dotenv = require('dotenv');
-const app = require('../config/app');
+const app = require('../../config/app');
 
 dotenv.config();
 
 describe('Test status endpoint.', () => {
   test('It should show service status.', (done) => {
     request(app)
-      .get('/status')
+      .get('/api/v1/status')
       .set('Accept', 'application/json')
       .then((response) => {
         expect(response.statusCode).toBe(200);
